refactor(update): split update.rs into per-panel input submodules
Extract the five panel key-handlers out of the 1131-line update.rs into
a dedicated src/input/ module, keeping update.rs focused on the TEA
update() loop and top-level key routing:

- src/input/mod.rs       — send_command() helper + pub re-exports
- src/input/watchlist.rs — handle_watchlist_key()
- src/input/positions.rs — handle_positions_key()
- src/input/orders.rs    — handle_orders_key()
- src/input/modal.rs     — handle_modal_key()
- src/input/search.rs    — handle_search_key()

src/update.rs now imports from crate::input and retains only update(),
handle_key(), handle_panel_key(), and the full test suite.

src/main.rs gains `mod input;` to declare the new binary-crate module.

README.md crate-structure diagram updated to reflect the new layout.

Closes #8
