+++ src/update.rs
@@ -1,27 +1,11 @@
 use crossterm::event::{KeyCode, KeyModifiers};
-use tokio::sync::mpsc::error::TrySendError;
 
-use crate::app::{App, ConfirmAction, Modal, OrderEntryState, OrderField, OrdersSubTab, Tab};
-use crate::commands::Command;
+use crate::app::{App, Modal, Tab};
 use crate::events::Event;
-
-/// Send a command on the command channel and set the appropriate status message.
-///
-/// - Success → `success_msg`
-/// - Channel full → "System busy — please retry"
-/// - Channel closed → "Command handler stopped — restart app" (+ error log)
-fn send_command(app: &mut App, cmd: Command, success_msg: impl Into<String>) {
-    match app.command_tx.try_send(cmd) {
-        Ok(()) => app.status_msg = success_msg.into(),
-        Err(TrySendError::Full(_)) => {
-            app.status_msg = "System busy — please retry".into();
-        }
-        Err(TrySendError::Closed(_)) => {
-            tracing::error!("command channel closed; command handler has stopped");
-            app.status_msg = "Command handler stopped — restart app".into();
-        }
-    }
-}
+use crate::input::{
+    handle_modal_key, handle_orders_key, handle_positions_key, handle_search_key,
+    handle_watchlist_key,
+};
 
 pub fn update(app: &mut App, event: Event) {
     match event {
@@ -116,352 +100,12 @@ fn handle_panel_key(app: &mut App, key: crossterm::event::KeyEvent) {
     }
 }
 
-fn handle_watchlist_key(app: &mut App, key: crossterm::event::KeyEvent) {
-    let len = app.watchlist.as_ref().map(|w| w.assets.len()).unwrap_or(0);
-
-    match key.code {
-        KeyCode::Char('j') | KeyCode::Down if len > 0 => {
-            let i = app.watchlist_state.selected().unwrap_or(0);
-            app.watchlist_state.select(Some((i + 1).min(len - 1)));
-        }
-        KeyCode::Char('k') | KeyCode::Up => {
-            let i = app.watchlist_state.selected().unwrap_or(0);
-            app.watchlist_state.select(Some(i.saturating_sub(1)));
-        }
-        KeyCode::Char('g') => app.watchlist_state.select(Some(0)),
-        KeyCode::Char('G') if len > 0 => {
-            app.watchlist_state.select(Some(len - 1));
-        }
-        KeyCode::Enter => {
-            if let Some(symbol) = app.selected_watchlist_symbol() {
-                app.modal = Some(Modal::SymbolDetail(symbol));
-            }
-        }
-        KeyCode::Char('o') => {
-            let symbol = app.selected_watchlist_symbol().unwrap_or_default();
-            app.modal = Some(Modal::OrderEntry(OrderEntryState::new(symbol)));
-        }
-        KeyCode::Char('a') => {
-            if let Some(wl) = &app.watchlist {
-                let id = wl.id.clone();
-                app.modal = Some(Modal::AddSymbol {
-                    input: String::new(),
-                    watchlist_id: id,
-                });
-            }
-        }
-        KeyCode::Char('d') => {
-            if let (Some(symbol), Some(wl)) =
-                (app.selected_watchlist_symbol(), app.watchlist.as_ref())
-            {
-                let wl_id = wl.id.clone();
-                app.modal = Some(Modal::Confirm {
-                    message: format!("Remove {} from watchlist?", symbol),
-                    action: ConfirmAction::RemoveFromWatchlist {
-                        watchlist_id: wl_id,
-                        symbol,
-                    },
-                    confirmed: false,
-                });
-            }
-        }
-        KeyCode::Char('/') => {
-            app.searching = true;
-            app.search_query.clear();
-        }
-        _ => {}
-    }
-}
-
-fn handle_positions_key(app: &mut App, key: crossterm::event::KeyEvent) {
-    let len = app.positions.len();
-    match key.code {
-        KeyCode::Char('j') | KeyCode::Down if len > 0 => {
-            let i = app.positions_state.selected().unwrap_or(0);
-            app.positions_state.select(Some((i + 1).min(len - 1)));
-        }
-        KeyCode::Char('k') | KeyCode::Up => {
-            let i = app.positions_state.selected().unwrap_or(0);
-            app.positions_state.select(Some(i.saturating_sub(1)));
-        }
-        KeyCode::Char('g') => app.positions_state.select(Some(0)),
-        KeyCode::Char('G') if len > 0 => {
-            app.positions_state.select(Some(len - 1));
-        }
-        KeyCode::Enter => {
-            if let Some(symbol) = app.selected_position_symbol() {
-                app.modal = Some(Modal::SymbolDetail(symbol));
-            }
-        }
-        KeyCode::Char('o') => {
-            let symbol = app.selected_position_symbol().unwrap_or_default();
-            let mut state = OrderEntryState::new(symbol);
-            state.side_buy = false;
-            app.modal = Some(Modal::OrderEntry(state));
-        }
-        _ => {}
-    }
-}
-
-fn handle_orders_key(app: &mut App, key: crossterm::event::KeyEvent) {
-    let orders = app.filtered_orders();
-    let len = orders.len();
-
-    match key.code {
-        KeyCode::Char('j') | KeyCode::Down if len > 0 => {
-            let i = app.orders_state.selected().unwrap_or(0);
-            app.orders_state.select(Some((i + 1).min(len - 1)));
-        }
-        KeyCode::Char('k') | KeyCode::Up => {
-            let i = app.orders_state.selected().unwrap_or(0);
-            app.orders_state.select(Some(i.saturating_sub(1)));
-        }
-        KeyCode::Char('g') => app.orders_state.select(Some(0)),
-        KeyCode::Char('G') if len > 0 => {
-            app.orders_state.select(Some(len - 1));
-        }
-        KeyCode::Char('1') => {
-            app.orders_subtab = OrdersSubTab::Open;
-            app.orders_state.select(Some(0));
-        }
-        KeyCode::Char('2') => {
-            app.orders_subtab = OrdersSubTab::Filled;
-            app.orders_state.select(Some(0));
-        }
-        KeyCode::Char('3') => {
-            app.orders_subtab = OrdersSubTab::Cancelled;
-            app.orders_state.select(Some(0));
-        }
-        KeyCode::Char('o') => {
-            app.modal = Some(Modal::OrderEntry(OrderEntryState::new(String::new())));
-        }
-        KeyCode::Char('c') => {
-            if let Some(id) = app.selected_order_id() {
-                app.modal = Some(Modal::Confirm {
-                    message: format!("Cancel order {}?", &id[..id.len().min(8)]),
-                    action: ConfirmAction::CancelOrder(id),
-                    confirmed: false,
-                });
-            }
-        }
-        _ => {}
-    }
-}
-
-fn handle_modal_key(app: &mut App, key: crossterm::event::KeyEvent) {
-    if key.code == KeyCode::Esc {
-        app.modal = None;
-        return;
-    }
-
-    // Clone to avoid borrow issues
-    let modal = match app.modal.take() {
-        Some(m) => m,
-        None => return,
-    };
-
-    let new_modal = match modal {
-        Modal::Help => {
-            if key.code != KeyCode::Esc {
-                None
-            } else {
-                Some(Modal::Help)
-            }
-        }
-
-        Modal::OrderEntry(mut state) => {
-            match key.code {
-                KeyCode::Tab => state.focused_field = state.focused_field.next(),
-                KeyCode::BackTab => state.focused_field = state.focused_field.prev(),
-                KeyCode::Left | KeyCode::Right => match state.focused_field {
-                    OrderField::Side => state.side_buy = !state.side_buy,
-                    OrderField::OrderType => state.market_order = !state.market_order,
-                    _ => {}
-                },
-                KeyCode::Char(c) => match state.focused_field {
-                    OrderField::Symbol => state.symbol.push(c),
-                    OrderField::Qty if c.is_ascii_digit() || c == '.' => {
-                        state.qty_input.push(c);
-                    }
-                    OrderField::Price if c.is_ascii_digit() || c == '.' => {
-                        state.price_input.push(c);
-                    }
-                    OrderField::Side => {
-                        if c == 'b' || c == 'B' {
-                            state.side_buy = true;
-                        } else if c == 's' || c == 'S' {
-                            state.side_buy = false;
-                        }
-                    }
-                    _ => {}
-                },
-                KeyCode::Backspace => match state.focused_field {
-                    OrderField::Symbol => {
-                        state.symbol.pop();
-                    }
-                    OrderField::Qty => {
-                        state.qty_input.pop();
-                    }
-                    OrderField::Price => {
-                        state.price_input.pop();
-                    }
-                    _ => {}
-                },
-                KeyCode::Enter => {
-                    if state.focused_field == OrderField::Submit {
-                        send_command(
-                            app,
-                            Command::SubmitOrder {
-                                symbol: state.symbol.clone(),
-                                side: if state.side_buy { "buy" } else { "sell" }.into(),
-                                order_type: if state.market_order {
-                                    "market"
-                                } else {
-                                    "limit"
-                                }
-                                .into(),
-                                qty: if state.qty_input.is_empty() {
-                                    None
-                                } else {
-                                    Some(state.qty_input.clone())
-                                },
-                                price: if state.market_order || state.price_input.is_empty() {
-                                    None
-                                } else {
-                                    Some(state.price_input.clone())
-                                },
-                                time_in_force: "day".into(),
-                            },
-                            "Submitting order…",
-                        );
-                        app.modal = None;
-                        return;
-                    } else {
-                        state.focused_field = state.focused_field.next();
-                    }
-                }
-                _ => {}
-            }
-            Some(Modal::OrderEntry(state))
-        }
-
-        Modal::SymbolDetail(_) => None,
-
-        Modal::Confirm {
-            message,
-            action,
-            mut confirmed,
-        } => match key.code {
-            KeyCode::Left | KeyCode::Right | KeyCode::Char('y') | KeyCode::Char('n') => {
-                confirmed = matches!(key.code, KeyCode::Char('y') | KeyCode::Left);
-                if confirmed {
-                    match &action {
-                        ConfirmAction::CancelOrder(id) => {
-                            send_command(
-                                app,
-                                Command::CancelOrder(id.clone()),
-                                format!("Cancelling {}…", &id[..id.len().min(8)]),
-                            );
-                        }
-                        ConfirmAction::RemoveFromWatchlist {
-                            watchlist_id,
-                            symbol,
-                        } => {
-                            send_command(
-                                app,
-                                Command::RemoveFromWatchlist {
-                                    watchlist_id: watchlist_id.clone(),
-                                    symbol: symbol.clone(),
-                                },
-                                format!("Removing {}…", symbol),
-                            );
-                        }
-                    }
-                    app.modal = None;
-                    return;
-                }
-                None
-            }
-            KeyCode::Enter => {
-                if confirmed {
-                    app.modal = None;
-                    return;
-                }
-                Some(Modal::Confirm {
-                    message,
-                    action,
-                    confirmed,
-                })
-            }
-            _ => Some(Modal::Confirm {
-                message,
-                action,
-                confirmed,
-            }),
-        },
-
-        Modal::AddSymbol {
-            mut input,
-            watchlist_id,
-        } => match key.code {
-            KeyCode::Char(c) => {
-                input.push(c.to_ascii_uppercase());
-                Some(Modal::AddSymbol {
-                    input,
-                    watchlist_id,
-                })
-            }
-            KeyCode::Backspace => {
-                input.pop();
-                Some(Modal::AddSymbol {
-                    input,
-                    watchlist_id,
-                })
-            }
-            KeyCode::Enter => {
-                if !input.is_empty() {
-                    send_command(
-                        app,
-                        Command::AddToWatchlist {
-                            watchlist_id: watchlist_id.clone(),
-                            symbol: input.clone(),
-                        },
-                        format!("Adding {}…", input),
-                    );
-                }
-                None
-            }
-            _ => Some(Modal::AddSymbol {
-                input,
-                watchlist_id,
-            }),
-        },
-    };
-
-    app.modal = new_modal;
-}
-
-fn handle_search_key(app: &mut App, key: crossterm::event::KeyEvent) {
-    match key.code {
-        KeyCode::Esc | KeyCode::Enter => {
-            app.searching = false;
-        }
-        KeyCode::Backspace => {
-            app.search_query.pop();
-        }
-        KeyCode::Char(c) => {
-            app.search_query.push(c);
-            app.watchlist_state.select(Some(0));
-        }
-        _ => {}
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::app::test_helpers::*;
     use crate::app::{Modal, OrdersSubTab, Tab};
+    use crate::commands::Command;
     use crate::events::Event;
     use crate::types::{AccountInfo, MarketClock, Order, Quote};
     use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
